Move software-model integ tests out of subprojects/core into plugins-model-native
Consolidates the remaining software-model integration tests that were
sitting under subprojects/core/src/integTest:

  - RuleTask{Bridging,Creation,Execution}IntegrationTest — wholesale moves;
    they exercise core's RuleSource-to-task bridging end-to-end.
  - NonImperativeBuildScriptEvaluationIntegrationTest — wholesale move;
    its only assertion is that model { } script plugins evaluate.
  - The "camel case software model tasks" test from
    TaskSelectionIntegrationTest and the "methods + model block" test from
    BuildScriptVisibilityIntegrationTest are extracted into a new
    SoftwareModelTaskAndBuildScriptIntegrationTest in plugins-model-native.
  - The two RuleSource-applying tests in PluginDetectionIntegrationTest
    are extracted into SoftwareModelPluginDetectionIntegrationTest.

After this commit, subprojects/core/src/integTest has zero `model {`
and zero `extends RuleSource` matches.
