+++ platforms/native/plugins-model-native/src/integTest/groovy/org/gradle/execution/taskgraph/RuleTaskExecutionIntegrationTest.groovy
