+++ platforms/native/plugins-model-native/src/integTest/groovy/org/gradle/plugin/NonImperativeBuildScriptEvaluationIntegrationTest.groovy
