+++ subprojects/core/src/integTest/groovy/org/gradle/api/PluginDetectionIntegrationTest.groovy
@@ -144,73 +144,6 @@ class PluginDetectionIntegrationTest extends AbstractIntegrationSpec {
         succeeds "help"
     }
 
-    def "plugin manager with id is fired after the plugin is applied for hybrid plugins"() {
-        when:
-        file("buildSrc/src/main/groovy/MyPlugin.groovy") << """
-            import org.gradle.api.Plugin
-            import org.gradle.api.Task
-            import org.gradle.model.*
-
-            class MyPlugin implements Plugin {
-                void apply(project) {
-                  project.tasks.create("imperative-sentinel")
-                }
-
-                static class Rules extends RuleSource {
-                    @Model String thing() {
-                        "foo"
-                    }
-                }
-            }
-        """
-
-        file("buildSrc/src/main/resources/META-INF/gradle-plugins/my.properties") << "implementation-class=MyPlugin"
-
-        buildFile """
-            import org.gradle.model.internal.core.ModelPath
-
-            pluginManager.withPlugin("my") {
-              assert tasks."imperative-sentinel"
-              // note: modelRegistry property is internal on project
-              assert modelRegistry.node(ModelPath.path("thing")) != null
-            }
-
-            pluginManager.apply(MyPlugin)
-        """
-
-        then:
-        succeeds "help"
-    }
-
-    def "plugin manager with id is fired after the plugin is applied for rule plugins"() {
-        when:
-        file("buildSrc/src/main/groovy/MyPlugin.groovy") << """
-            import org.gradle.model.*
-
-            class Rules extends RuleSource {
-                @Model String thing() {
-                    "foo"
-                }
-            }
-        """
-
-        file("buildSrc/src/main/resources/META-INF/gradle-plugins/my.properties") << "implementation-class=Rules"
-
-        buildFile """
-            import org.gradle.model.internal.core.ModelPath
-
-            pluginManager.withPlugin("my") {
-              // note: modelRegistry property is internal on project
-              assert modelRegistry.node(ModelPath.path("thing")) != null
-            }
-
-            pluginManager.apply("my")
-        """
-
-        then:
-        succeeds "help"
-    }
-
     @Issue("http://discuss.gradle.org/t/concurrentmodification-exception-on-java-8-for-plugins-withid-with-gradle-2-4/8928")
     def "can nest detection"() {
         // Actual plugins in use here are insignificant
