+++ subprojects/core/src/integTest/groovy/org/gradle/api/tasks/TaskSelectionIntegrationTest.groovy
@@ -106,22 +106,6 @@ allprojects { task thing }
         result.assertTasksScheduled(":child:child:thing")
     }
 
-    def "can use camel case to match software model tasks"() {
-        buildFile << """
-            model {
-                tasks {
-                    "sayHelloToUser"(DefaultTask) {
-                    }
-                }
-            }
-        """
-        when:
-        run "sHTU"
-
-        then:
-        result.assertTasksScheduled(":sayHelloToUser")
-    }
-
     @ToBeFixedForIsolatedProjects(because = "allprojects")
     def "executes project default tasks when none specified"() {
         createDirs("a")
